Implemented an alternative connection type for self debugging purposes - Decoupled `CSteamId` from identifying an online player, now using generated id - Both steam users and single non-steam user can join a lobby, although mixing the two are not supported properly - Respawning is not working properly with steam connection - Disconnecting using UDPClient is not fully implemented - Modified project files

+++ Online/OnlineEntity.EntityId.cs
@@ -4,11 +4,11 @@
     {
         public class EntityId : System.IEquatable<EntityId>, Serializer.ICustomSerializable // How we refer to a game entity online
         {
-            public ulong originalOwner;
+            public int originalOwner;
             public int id;
 
             public EntityId() { }
-            public EntityId(ulong originalOwner, int id)
+            public EntityId(int originalOwner, int id)
             {
                 this.originalOwner = originalOwner;
                 this.id = id;
